DLS-8992 Move ConfigurationOps to their own file
